refactor: expose all services via mrtd package index
